requirements.txt
pymediainfo>=6.0.0
guessit>=3.8.0
rich>=13.0.0
requests>=2.28.0
python-dotenv>=1.0.0
